requirements.txt
python-telegram-bot
openai== 0.27.6
python-telegram-bot
requests
sentence_transformers
einops